Add `use_bundler` configuration option (#19)
Adds a new configuration option called `use_bundler` that will indicate
whether to start an LSP server with the `bundle exec` command.

Priority list (top to bottom):

1. An absolute path from `lsp.<LSP_NAME>.binary.path`. Ignore the
`use_bundler` configuration option value. No fallback to the option
below.
2. If the `use_bundler` configuration option is **true**:
- Check if `bundle` is available in the **PATH**. In 99% of cases, this
should be true.
- Start the LSP using the following command: `bundle exec <LSP_NAME>`.
3. If the `use_bundler` configuration option is **false**:
- Use the current logic by checking the **PATH** for the LSP executable,
i.e., `which ruby-lsp`. Most of the time, this option expects to find an
executable as a binstub in the `<project_dir>/bin` directory.