Use bundler for `solargraph` and `rubocop` (#22)
That improves the developer experience (DX) when using `solargraph` and 
`rubocop` LSPs. Both LSPs work best when started in the context of
Bundler.

**How does it work currently?** Here is an example with `rubocop`: if 
`rubocop` exists in the PATH, use it. So end users have two options:
specify
the full path to the `rubocop` binary (or to `bundle`) or generate a
binstub
and then add the `<project_dir>/bin` folder to PATH.

This is not optimal because most of the time we want to use and spawn
LSPs in
the context of Bundler to ensure we load the correct gemset. This does
not
fully apply to `ruby-lsp`, which can work without `bundle exec`.

Overall, the new configuration option `use_bundler` is configured as
follows:

1. **Solargraph:** Use Bundler by default.
2. **Rubocop:** Use Bundler by default.
3. **Ruby LSP:** Do not use Bundler by default.