Update `herb-language-server` executable path (#115)
Just a small pull request to update the `herb-language-server`
executable as discussed on Discord.

In [v0.3.1](https://github.com/marcoroth/herb/releases/tag/v0.3.1) we
changed the location of the `herb-language-server` executable from
`./dist/herb-language-server` to `./bin/herb-language-server` to be more
flexible if we would want to swap the implementation in the future.

This shouldn't be a breaking change since #110 hasn't been released yet.