Ensure cwd actually has a default (#127)
Turns out, the default cwd was getting set on the JSON config after we
had already deserialized it into a RubyDebugConfig. This meant there was
not in fact a default cwd, so if a config failed to specify, the launch
would fail with a cryptic:
>process exited before debugger could connect
This is because bundle/rdbg would immediately exit, as it was being run
from the home directory, not the project's directory.

closes #126