fix(gemset): pass `GEM_PATH` env variable to commands
On Linux, for some reason, it's not enough
to set `GEM_HOME` env variable to isolate the
extension gemset:

```
  - GEM PATHS:
     -
/home/vslobodin/.local/share/mise/installs/ruby/3.4.5/lib/ruby/gems/3.4.0
     - /home/vslobodin/.local/share/zed/extensions/work/ruby
```

The above snippet is the result from running `mise` and `gem env`.
User installed gemset is taken into account
which breaks installation of extension gems.
Setting the `GEM_PATH` env var seems to fix that issue.