Add Zed extension workflows (#214)
Hey everyone! 

We recently started working on dedicated workflows for extensions and
this is the result of this!

I'd like to add them here sooner than later (we eventually would like to
set up org-wide rulesets for these), but I see this conflicts currently
with your workflows already and I definitely do not want to remove your
setup completely. Do we want to keep the jobs for the styling and merge
it then?

Happy to hear your opinion on this, feel free to reach out with any
other feedback also!